# api/openapi/mvp.yaml
openapi: 3.1.0
info:
  title: YOLO Platform MVP API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: OK
  /readyz:
    get:
      summary: Readiness probe
      responses:
        "200":
          description: Ready
  /v1/datasets:
    post:
      summary: Create dataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, name, bucket]
      responses:
        "200":
          description: Dataset created
  /v1/datasets/{id}/scan:
    post:
      summary: Scan dataset object keys
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan accepted
  /v1/datasets/{id}/snapshots:
    post:
      summary: Create snapshot
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Snapshot created
    get:
      summary: List snapshots
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Snapshot list
  /v1/snapshots/{id}/import:
    post:
      summary: Queue snapshot import job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format, idempotency_key]
              properties:
                format:
                  type: string
                  enum: [yolo, coco]
      responses:
        "200":
          description: Import job accepted
  /v1/snapshots/{id}/export:
    post:
      summary: Queue snapshot export package build (currently yolo only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export job accepted
  /v1/datasets/{id}/items:
    get:
      summary: List indexed dataset items
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item list
  /v1/objects/presign:
    post:
      summary: Create object read presign URL
      responses:
        "200":
          description: Signed URL
  /v1/projects/{id}/overview:
    get:
      summary: Get project task overview
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project overview payload
  /v1/projects/{id}/tasks:
    get:
      summary: List project tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: kind
          schema:
            type: string
        - in: query
          name: assignee
          schema:
            type: string
        - in: query
          name: priority
          schema:
            type: string
        - in: query
          name: snapshot_id
          schema:
            type: integer
      responses:
        "200":
          description: Project task list
    post:
      summary: Create project task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
      responses:
        "201":
          description: Task created
  /v1/tasks/{id}:
    get:
      summary: Get task detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task detail payload
  /v1/tasks/{id}/transition:
    post:
      summary: Transition task status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
      responses:
        "200":
          description: Updated task payload
  /v1/tasks/{id}/workspace:
    get:
      summary: Get annotation workspace payload
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Annotation workspace payload
  /v1/tasks/{id}/workspace/draft:
    put:
      summary: Save annotation workspace draft
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actor, body]
      responses:
        "200":
          description: Updated workspace payload
  /v1/tasks/{id}/workspace/submit:
    post:
      summary: Submit annotation workspace task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actor]
      responses:
        "200":
          description: Submitted workspace payload
  /v1/jobs/zero-shot:
    post:
      summary: Queue zero-shot job
      responses:
        "200":
          description: Job accepted
  /v1/jobs/video-extract:
    post:
      summary: Queue video extraction job
      responses:
        "200":
          description: Job accepted
  /v1/jobs/cleaning:
    post:
      summary: Queue cleaning job
      responses:
        "200":
          description: Job accepted
  /v1/jobs/{job_id}:
    get:
      summary: Get job status
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
  /v1/jobs/{job_id}/events:
    get:
      summary: Get job event timeline
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job event list
  /v1/snapshots/diff:
    post:
      summary: Diff snapshots
      responses:
        "200":
          description: Diff result
  /v1/review/candidates:
    get:
      summary: List queued review candidates with source metadata
      responses:
        "200":
          description: Candidate list including lifecycle status, confidence, source model, and provenance metadata
  /v1/review/candidates/{id}/accept:
    post:
      summary: Accept candidate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate accepted
  /v1/review/candidates/{id}/reject:
    post:
      summary: Reject candidate with required reason code
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason_code
              properties:
                reviewer_id:
                  type: string
                reason_code:
                  type: string
      responses:
        "200":
          description: Candidate rejected
  /v1/publish/candidates:
    get:
      summary: List publish candidate groups
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Publish candidate groups
  /v1/publish/batches:
    post:
      summary: Create publish batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, snapshot_id, source, items]
      responses:
        "201":
          description: Publish batch created
  /v1/publish/batches/{id}:
    get:
      summary: Get publish batch detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Publish batch detail
  /v1/publish/batches/{id}/items:
    post:
      summary: Replace publish batch items
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Publish batch items replaced
  /v1/publish/batches/{id}/review-approve:
    post:
      summary: Approve publish batch as reviewer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviewer approval recorded
  /v1/publish/batches/{id}/review-reject:
    post:
      summary: Reject publish batch as reviewer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviewer rejection recorded
  /v1/publish/batches/{id}/review-rework:
    post:
      summary: Request publish batch rework as reviewer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviewer rework recorded
  /v1/publish/batches/{id}/owner-approve:
    post:
      summary: Approve publish batch as owner
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Owner approval recorded
  /v1/publish/batches/{id}/owner-reject:
    post:
      summary: Reject publish batch as owner
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Owner rejection recorded
  /v1/publish/batches/{id}/owner-rework:
    post:
      summary: Request publish batch rework as owner
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Owner rework recorded
  /v1/publish/batches/{id}/feedback:
    post:
      summary: Add batch-level publish feedback
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Publish batch feedback created
  /v1/publish/batches/{id}/items/{itemId}/feedback:
    post:
      summary: Add item-level publish feedback
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Publish item feedback created
  /v1/publish/batches/{id}/workspace:
    get:
      summary: Get aggregated review workspace payload
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review workspace payload
  /v1/publish/records/{id}:
    get:
      summary: Get publish record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Publish record detail
  /v1/artifacts/packages:
    post:
      summary: Queue artifact package build (currently yolo only)
      responses:
        "202":
          description: Package job accepted
        "400":
          description: Invalid package request
  /v1/artifacts/resolve:
    get:
      summary: Resolve artifact metadata by optional dataset, format, and version
      parameters:
        - in: query
          name: dataset
          required: false
          schema:
            type: string
        - in: query
          name: format
          required: true
          schema:
            type: string
        - in: query
          name: version
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved artifact metadata
        "400":
          description: Missing required query parameters
        "404":
          description: Ready artifact not found
  /v1/artifacts/{id}:
    get:
      summary: Get artifact metadata and inline bundle entries
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artifact details
        "404":
          description: Artifact not found
  /v1/artifacts/{id}/download:
    get:
      summary: Download artifact package archive
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artifact archive stream
        "404":
          description: Artifact not found
        "409":
          description: Artifact exists but is not ready for download
  /v1/artifacts/{id}/presign:
    post:
      summary: Create artifact download presign URL
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signed artifact URL
        "404":
          description: Artifact not found
        "409":
          description: Artifact exists but is not ready for download